Added missing methods
